Split 2 different uses of MUI Badge in 2 different components
